% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDOLOptimizedCpGs450klegacy.R
\docType{data}
\name{IDOLOptimizedCpGs450klegacy}
\alias{IDOLOptimizedCpGs450klegacy}
\title{IDOL Optimized CpGs for adult blood DNA methylation deconvolution EPIC}
\format{An object of class "character" of length 350.

        The format is:
        chr [1:350] "cg14232368" "cg15087459" "cg20538211" "cg11944101" "cg00219921" "cg23752651" "cg10978799" "cg22442430" "cg12294633" "cg05492387" "cg24336674" "cg25939861" ...}
\usage{
IDOLOptimizedCpGs450klegacy
}
\description{
This object is nothing more than a vector of length 350 consisting of the
    names of the IDOL optimized CpGs to use in the older 450K platform.  
    These CpGs are used as the backbone for deconvolution and were selected 
    because their methylation signature differs across the six
    normal leukocyte subtypes.
}
\examples{
# Do not run
# data("IDOLOptimizedCpGs450klegacy")
}
\references{
D Koestler et al. (2016). \emph{Improving cell mixture 
deconvolution by identifying optimal DNA methylation libraries (IDOL)}. 
BMC bioinformatics. 17, 120.
}
\keyword{datasets}
