% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateCellCounts2.R
\name{estimateCellCounts2}
\alias{estimateCellCounts2}
\title{estimateCellCounts2 function allows the use of customized reference datasets 
and IDOL probes L-DMR lists}
\usage{
estimateCellCounts2(rgSet, compositeCellType = "Blood",
  processMethod = c("auto", "preprocessNoob"), probeSelect = c("auto",
  "any", "IDOL"), cellTypes = c("CD8T", "CD4T", "NK", "Bcell", "Mono", "Neu"),
  referencePlatform = c("IlluminaHumanMethylation450k",
  "IlluminaHumanMethylationEPIC", "IlluminaHumanMethylation27k"),
  referenceset = NULL, IDOLOptimizedCpGs = NULL, returnAll = FALSE,
  meanPlot = FALSE, verbose = TRUE, ...)
}
\description{
estimateCellCounts2 function allows the use of customized reference datasets 
and IDOL probes L-DMR lists
}
\examples{
# Step 1: Load the library(FlowSorted.Blood.EPIC)
data(FlowSorted.Blood.EPIC)
# Step 2 separate the reference from the testing dataset if you want to run 
# examples for estimations for this function example
cell.Mix <- which(FlowSorted.Blood.EPIC$CellType == "MIX")
RGsetTargets <- FlowSorted.Blood.EPIC[, cell.Mix]
sampleNames(RGsetTargets) <- paste(RGsetTargets$CellType,
                            seq(along = cell.Mix), sep = "_")
# Step 3: use your favorite package for deconvolution.
# You can also read in the IDOL optimized DMR library based on the EPIC 
# array.  Thisobject is nothing more than a vector of length 450 consisting 
# of the names of the IDOL optimized CpGs.  These CpGs are used as the 
# backbone for deconvolution and were selected because their methylation 
# signature differs across the six normal leukocyte subtypes.
data(IDOLOptimizedCpGs)
# Step 4: Deconvolute a target data set consisting of EPIC DNA methlation 
# data profiled in blood, using your prefered method estimateCellCounts 
# (minfi), or similar.
# We recommend using Noob processMethod = "preprocessNoob" in minfi
# for the target and reference datasets Cell types included are 
# cellTypes = c("CD8T", "CD4T", "NK", "Bcell", "Mono", "Neu")
# We also provide an IDOL optimized list of CpGs (IDOLOptimizedCpGs)
# that can be used for optimized cell estimations
# We offer an adaptation of the popular estimateCellCounts in minfi to allow 
# the inclusion of customized reference arrays to select the custom IDOL use 
# probeSelect = "IDOL"
countsEPIC<-estimateCellCounts2(RGsetTargets, compositeCellType = "Blood", 
                                processMethod = "preprocessNoob",
                                probeSelect = "IDOL", 
                                cellTypes = c("CD8T", "CD4T", "NK", "Bcell", 
                                "Mono", "Neu"), 
                                referencePlatform = 
                                "IlluminaHumanMethylationEPIC",
                                referenceset=FlowSorted.Blood.EPIC, 
                                IDOLOptimizedCpGs =IDOLOptimizedCpGs, 
                                returnAll = FALSE)

}
