% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraryDataGet.R
\name{libraryDataGet}
\alias{libraryDataGet}
\title{libraryDataGet}
\usage{
libraryDataGet(title)
}
\arguments{
\item{title}{title of the data, e.g., 'FlowSorted.Blood.EPIC'}
}
\value{
The function will look for the dataset in ExperimentHub and load the object

This function will return an object matching the title of the ExperimenHub
}
\description{
Function to load the library data from ExperimentHub
}
\examples{
FlowSorted.Blood.EPIC<-
libraryDataGet('FlowSorted.Blood.EPIC')
FlowSorted.Blood.EPIC
}
